Commit from 20/7

+++ file_io/0-read_textfile.c
@@ -12,7 +12,7 @@
 ssize_t read_textfile(const char *filename, size_t letters)
 {
 	char *fptr = malloc(letters);
-	size_t ofile, rfile, wfile;
+	int ofile, rfile, wfile;
 
 	if (!fptr)
 	{
